Separate client sdk and consolidate the training_client
