+++ scripts/xml-rpc-client-example/training_client_example.py
@@ -1,188 +1,199 @@
-import xmlrpc.client
-import sys
-import json
-import time
+# import xmlrpc.client
+# import sys
+# import json
+# import time
 import os
-from datasets import load_dataset
-from pprint import pprint
 from datetime import datetime
-import logging
-from logging.handlers import RotatingFileHandler
+from pprint import pprint
+
+from datasets import load_dataset
+from tlab_sdk_client.callbacks.hf_callback import TLabProgressCallback
+from tlab_sdk_client.client import TransformerLabClient
+from transformers import (
+    AutoModelForCausalLM,
+    AutoTokenizer,
+    DataCollatorForLanguageModeling,
+    Trainer,
+    TrainingArguments,
+)
+
+# import logging
+# from logging.handlers import RotatingFileHandler
 
 
-class TransformerLabClient:
-    """Client for reporting training progress to TransformerLab via XML-RPC"""
+# class TransformerLabClient:
+#     """Client for reporting training progress to TransformerLab via XML-RPC"""
 
-    def __init__(self, server_url="http://localhost:8338/trainer_rpc", log_file=None):
-        """Initialize the XML-RPC client"""
-        self.server = xmlrpc.client.ServerProxy(server_url)
-        self.job_id = None
-        self.config = {}
-        self.last_report_time = 0
-        self.report_interval = 1  # seconds
-        self.log_file = log_file
+#     def __init__(self, server_url="http://localhost:8338/trainer_rpc", log_file=None):
+#         """Initialize the XML-RPC client"""
+#         self.server = xmlrpc.client.ServerProxy(server_url)
+#         self.job_id = None
+#         self.config = {}
+#         self.last_report_time = 0
+#         self.report_interval = 1  # seconds
+#         self.log_file = log_file
 
-    def start_job(self, config):
-        """Register job with TransformerLab and get a job ID"""
-        result = self.server.start_training(json.dumps(config))
-        if result["status"] == "started":
-            self.job_id = result["job_id"]
-            self.config = config
-            # Set up logger
-            self.create_logger(log_file=self.log_file)
-            self.log_info(f"Registered job with TransformerLab. Job ID: {self.job_id}")
-            return self.job_id
-        else:
-            error_msg = f"Failed to start job: {result['message']}"
-            raise Exception(error_msg)
+#     def start_job(self, config):
+#         """Register job with TransformerLab and get a job ID"""
+#         result = self.server.start_training(json.dumps(config))
+#         if result["status"] == "started":
+#             self.job_id = result["job_id"]
+#             self.config = config
+#             # Set up logger
+#             self.create_logger(log_file=self.log_file)
+#             self.log_info(f"Registered job with TransformerLab. Job ID: {self.job_id}")
+#             return self.job_id
+#         else:
+#             error_msg = f"Failed to start job: {result['message']}"
+#             raise Exception(error_msg)
 
-    def report_progress(self, progress, metrics=None):
-        """Report training progress to TransformerLab"""
-        if not self.job_id:
-            return True
+#     def report_progress(self, progress, metrics=None):
+#         """Report training progress to TransformerLab"""
+#         if not self.job_id:
+#             return True
 
-        # Rate limit reports
-        current_time = time.time()
-        if current_time - self.last_report_time < self.report_interval:
-            return True
+#         # Rate limit reports
+#         current_time = time.time()
+#         if current_time - self.last_report_time < self.report_interval:
+#             return True
 
-        self.last_report_time = current_time
+#         self.last_report_time = current_time
 
-        try:
-            status = self.server.get_training_status(self.job_id, int(progress))
+#         try:
+#             status = self.server.get_training_status(self.job_id, int(progress))
 
-            # If metrics are important, consider logging them separately
-            if metrics and hasattr(self.server, "log_metrics"):
-                self.server.log_metrics(self.job_id, json.dumps(metrics))
+#             # If metrics are important, consider logging them separately
+#             if metrics and hasattr(self.server, "log_metrics"):
+#                 self.server.log_metrics(self.job_id, json.dumps(metrics))
 
-            if status.get("status") == "stopped":
-                self.log_info("Job was stopped remotely. Terminating training...")
-                sys.exit(1)
-                return False
-            return True
-        except Exception as e:
-            print(f"Error reporting progress: {e}")
-            # Still return True to continue training despite reporting error
-            return True
+#             if status.get("status") == "stopped":
+#                 self.log_info("Job was stopped remotely. Terminating training...")
+#                 sys.exit(1)
+#                 return False
+#             return True
+#         except Exception as e:
+#             print(f"Error reporting progress: {e}")
+#             # Still return True to continue training despite reporting error
+#             return True
 
-    def complete_job(self, message="Training completed successfully"):
-        """Mark job as complete in TransformerLab"""
-        if not self.job_id:
-            return
+#     def complete_job(self, message="Training completed successfully"):
+#         """Mark job as complete in TransformerLab"""
+#         if not self.job_id:
+#             return
 
-        try:
-            # Use the dedicated complete_job method if it exists
-            if hasattr(self.server, "complete_job"):
-                self.server.complete_job(self.job_id, "COMPLETE", message)
-            else:
-                # Fall back to using get_training_status with 100% progress
-                self.report_progress(100)
-                self.server.get_training_status(self.job_id, 100)
-        except Exception as e:
-            self.log_error(f"Error completing job: {e}")
+#         try:
+#             # Use the dedicated complete_job method if it exists
+#             if hasattr(self.server, "complete_job"):
+#                 self.server.complete_job(self.job_id, "COMPLETE", message)
+#             else:
+#                 # Fall back to using get_training_status with 100% progress
+#                 self.report_progress(100)
+#                 self.server.get_training_status(self.job_id, 100)
+#         except Exception as e:
+#             self.log_error(f"Error completing job: {e}")
 
-    def stop_job(self, message="Training completed successfully"):
-        """Mark job as complete in TransformerLab"""
-        if not self.job_id:
-            return
+#     def stop_job(self, message="Training completed successfully"):
+#         """Mark job as complete in TransformerLab"""
+#         if not self.job_id:
+#             return
 
-        try:
-            # Use the dedicated complete_job method if it exists
-            if hasattr(self.server, "complete_job"):
-                self.server.complete_job(self.job_id, "STOPPED", message)
-            else:
-                # Fall back to using get_training_status with 100% progress
-                self.report_progress(100)
-                self.server.get_training_status(self.job_id, 100)
-        except Exception as e:
-            self.log_error(f"Error completing job: {e}")
+#         try:
+#             # Use the dedicated complete_job method if it exists
+#             if hasattr(self.server, "complete_job"):
+#                 self.server.complete_job(self.job_id, "STOPPED", message)
+#             else:
+#                 # Fall back to using get_training_status with 100% progress
+#                 self.report_progress(100)
+#                 self.server.get_training_status(self.job_id, 100)
+#         except Exception as e:
+#             self.log_error(f"Error completing job: {e}")
 
-    def save_model(self, saved_model_path: str):
-        """Save the model to the specified path"""
-        if not self.job_id:
-            return
+#     def save_model(self, saved_model_path: str):
+#         """Save the model to the specified path"""
+#         if not self.job_id:
+#             return
 
-        try:
-            # Use the dedicated save_model method if it exists
-            if hasattr(self.server, "save_model"):
-                response = self.server.save_model(self.job_id, os.path.abspath(saved_model_path))
-                print("RESPONSE", response)
+#         try:
+#             # Use the dedicated save_model method if it exists
+#             if hasattr(self.server, "save_model"):
+#                 response = self.server.save_model(self.job_id, os.path.abspath(saved_model_path))
+#                 print("RESPONSE", response)
 
-            else:
-                self.log_warning("save_model method not available in server.")
-        except Exception as e:
-            print(f"Error saving model: {e}")
+#             else:
+#                 self.log_warning("save_model method not available in server.")
+#         except Exception as e:
+#             print(f"Error saving model: {e}")
 
-    def update_output_file_in_tlab(self):
-        try:
-            if hasattr(self.server, "update_output_file"):
-                # Use the dedicated update_output_file method if it exists
-                self.server.update_output_file(self.job_id, os.path.abspath(self.log_file_path))
-            else:
-                print("There was an issue with updating output.txt within Transformer Lab app.")
-        except Exception as e:
-            print(f"There was an issue with updating output.txt within Transformer Lab app: {str(e)}")
-            raise e
+#     def update_output_file_in_tlab(self):
+#         try:
+#             if hasattr(self.server, "update_output_file"):
+#                 # Use the dedicated update_output_file method if it exists
+#                 self.server.update_output_file(self.job_id, os.path.abspath(self.log_file_path))
+#             else:
+#                 print("There was an issue with updating output.txt within Transformer Lab app.")
+#         except Exception as e:
+#             print(f"There was an issue with updating output.txt within Transformer Lab app: {str(e)}")
+#             raise e
 
-    def create_logger(self, log_file=None, level=logging.INFO):
-        """Initialize logger with both file and console output"""
-        # Create logger
-        self.logger = logging.getLogger("transformerlab")
-        self.logger.setLevel(level)
-        self.logger.handlers = []  # Clear any existing handlers
+#     def create_logger(self, log_file=None, level=logging.INFO):
+#         """Initialize logger with both file and console output"""
+#         # Create logger
+#         self.logger = logging.getLogger("transformerlab")
+#         self.logger.setLevel(level)
+#         self.logger.handlers = []  # Clear any existing handlers
 
-        # Create formatter
-        formatter = logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
+#         # Create formatter
+#         formatter = logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
 
-        # If no log file specified, create one with timestamp
-        if not log_file:
-            os.makedirs("logs", exist_ok=True)
-            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-            log_file = f"logs/transformerlab_training_{timestamp}.log"
+#         # If no log file specified, create one with timestamp
+#         if not log_file:
+#             os.makedirs("logs", exist_ok=True)
+#             timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+#             log_file = f"logs/transformerlab_training_{timestamp}.log"
 
-        self.log_file_path = log_file
+#         self.log_file_path = log_file
 
-        # Create file handler
-        file_handler = RotatingFileHandler(
-            log_file,
-            maxBytes=10 * 1024 * 1024,  # 10MB
-            backupCount=5,
-        )
-        file_handler.setFormatter(formatter)
-        self.logger.addHandler(file_handler)
+#         # Create file handler
+#         file_handler = RotatingFileHandler(
+#             log_file,
+#             maxBytes=10 * 1024 * 1024,  # 10MB
+#             backupCount=5,
+#         )
+#         file_handler.setFormatter(formatter)
+#         self.logger.addHandler(file_handler)
 
-        # Create console handler
-        console_handler = logging.StreamHandler(sys.stdout)
-        console_handler.setFormatter(formatter)
-        self.logger.addHandler(console_handler)
+#         # Create console handler
+#         console_handler = logging.StreamHandler(sys.stdout)
+#         console_handler.setFormatter(formatter)
+#         self.logger.addHandler(console_handler)
 
-        self.log_file = log_file
-        self.log_info(f"Logging to file: {log_file}")
+#         self.log_file = log_file
+#         self.log_info(f"Logging to file: {log_file}")
 
-    def log_info(self, message):
-        """Log info message"""
-        self.logger.info(message)
-        self.update_output_file_in_tlab()
+#     def log_info(self, message):
+#         """Log info message"""
+#         self.logger.info(message)
+#         self.update_output_file_in_tlab()
 
-    def log_error(self, message):
-        """Log error message"""
-        self.logger.error(message)
-        self.update_output_file_in_tlab()
+#     def log_error(self, message):
+#         """Log error message"""
+#         self.logger.error(message)
+#         self.update_output_file_in_tlab()
 
-    def log_warning(self, message):
-        """Log warning message"""
-        self.logger.warning(message)
-        self.update_output_file_in_tlab()
+#     def log_warning(self, message):
+#         """Log warning message"""
+#         self.logger.warning(message)
+#         self.update_output_file_in_tlab()
 
-    def log_debug(self, message):
-        """Log debug message"""
-        self.logger.debug(message)
-        self.update_output_file_in_tlab()
+#     def log_debug(self, message):
+#         """Log debug message"""
+#         self.logger.debug(message)
+#         self.update_output_file_in_tlab()
 
-    def log_critical(self, message):
-        """Log critical message"""
-        self.logger.critical(message)
-        self.update_output_file_in_tlab()
+#     def log_critical(self, message):
+#         """Log critical message"""
+#         self.logger.critical(message)
+#         self.update_output_file_in_tlab()
 
 
 def train():
@@ -212,7 +223,7 @@ def train():
     tlab_client = TransformerLabClient()
     job_id = tlab_client.start_job(training_config)
 
-    # Create output directory
+    # Create output directory if it doesn't exist
     os.makedirs(training_config["output_dir"], exist_ok=True)
 
     try:
@@ -229,7 +240,6 @@ def train():
         tlab_client.report_progress(10, {"status": "dataset_loaded"})
 
         # Load tokenizer and model
-        from transformers import AutoTokenizer, AutoModelForCausalLM
 
         tlab_client.log_info(f"Loading model: {training_config['model_name']}")
         tokenizer = AutoTokenizer.from_pretrained(training_config["model_name"])
@@ -280,7 +290,6 @@ def train():
         tlab_client.report_progress(30, {"status": "dataset_processed"})
 
         # Setup training arguments
-        from transformers import TrainingArguments, Trainer, DataCollatorForLanguageModeling
 
         training_args = TrainingArguments(
             output_dir=os.path.join(training_config["output_dir"], f"job_{job_id}"),
@@ -297,34 +306,34 @@ def train():
         )
 
         # Define a custom callback to report progress to TransformerLab
-        from transformers import TrainerCallback
+        # from transformers import TrainerCallback
 
-        class TLabProgressCallback(TrainerCallback):
-            def __init__(self, tlab_client):
-                self.tlab_client = tlab_client
+        # class TLabProgressCallback(TrainerCallback):
+        #     def __init__(self, tlab_client):
+        #         self.tlab_client = tlab_client
 
-            def on_step_end(self, args, state, control, **kwargs):
-                if state.is_local_process_zero:
-                    if state.max_steps > 0:
-                        # Calculate progress percentage (30-90%)
-                        progress = 30 + ((state.global_step / state.max_steps) * 60)
-                        metrics = {
-                            "step": state.global_step,
-                            "train/loss": state.log_history[-1]["loss"] if state.log_history else None,
-                        }
-                        # Report progress to TransformerLab
-                        if not self.tlab_client.report_progress(progress, metrics):
-                            # Job was stopped remotely
-                            control.should_training_stop = True
+        #     def on_step_end(self, args, state, control, **kwargs):
+        #         if state.is_local_process_zero:
+        #             if state.max_steps > 0:
+        #                 # Calculate progress percentage (30-90%)
+        #                 progress = 30 + ((state.global_step / state.max_steps) * 60)
+        #                 metrics = {
+        #                     "step": state.global_step,
+        #                     "train/loss": state.log_history[-1]["loss"] if state.log_history else None,
+        #                 }
+        #                 # Report progress to TransformerLab
+        #                 if not self.tlab_client.report_progress(progress, metrics):
+        #                     # Job was stopped remotely
+        #                     control.should_training_stop = True
 
-            def on_log(self, args, state, control, logs=None, **kwargs):
-                if logs and "loss" in logs:
-                    metrics = {"step": state.global_step, "train/loss": logs["loss"]}
-                    # Add other metrics if available
-                    for key, value in logs.items():
-                        if isinstance(value, (int, float)):
-                            metrics[key] = value
-                    self.tlab_client.report_progress(30 + ((state.global_step / state.max_steps) * 60), metrics)
+        #     def on_log(self, args, state, control, logs=None, **kwargs):
+        #         if logs and "loss" in logs:
+        #             metrics = {"step": state.global_step, "train/loss": logs["loss"]}
+        #             # Add other metrics if available
+        #             for key, value in logs.items():
+        #                 if isinstance(value, (int, float)):
+        #                     metrics[key] = value
+        #             self.tlab_client.report_progress(30 + ((state.global_step / state.max_steps) * 60), metrics)
 
         # Setup trainer
         trainer = Trainer(
